Create README.md
Make a program which will allow the user to tag/highlight data that are unique and significant enough to undergo extensive analyses to enhance or improve autonomous drive algorithms, using Wi-Fi compunction through The Transmission Control Protocol (TCP) protocol. TCP is a core protocol of the Internet protocol suite. It originated in the initial network implementation in which it complemented the Internet Protocol (IP). Therefore, the entire suite is commonly referred to as TCP/IP. TCP provides reliable, ordered, and error-checked delivery of a stream of octets between applications running on hosts communicating over an IP network. AD interface is fast and efficient access to large data base and quickly retrieve desired scenarios and feature for analyse. In addition, it will help to reduce the time required to analyse data. 

♣	Program Requirements
ϖ	Two Applications have to be running at the same time in order to achieve a success communication between the client (IOS device) and the server (PC device). 
ϖ	Both Applications have to have the same IP Address and same port number order to achieve a success communication between the client (IOS device) and the server (PC device). 
ϖ	The client (IOS device) have to be able to send data to the server (PC device) using TCP protocol. 
ϖ	The server (PC device) have to be able to receive data from the client (IOS device) using TCP protocol.
ϖ	In the PC device, the user should be able to store these data inside a text file as table. 
ϖ	In the IOS device, the user should be able to see error checking if there’s no communication with the server to avoid losing data once tagged. 
ϖ	In the IOS device, the user should be able to navigate very quickly between these pages either by swiping or by clicking the buttons. 
ϖ	In the IOS device, the user should be able to see the tab bar color changed to blue if there’s communication with the server and to red if there’s no communication with the server. In addition, the user should be able to see a label at the top of the screen that show if there’s connection.